Implement agent-worker claim/heartbeat loop; extract @mymemo/agent-db (#180)
Task 3.1: the agent-worker control loop over the shared run-store helpers.
Each tick heartbeats owned runs (renew locked_until, observe cancel_requested
or ownership loss) then claims queued runs up to the supervisor's capacity,
processes each (Milestone 3: a synthetic one-text-event turn), and
terminalizes with the right outcome (done / error / canceled).

To let both chat-api and agent-worker run ONE implementation of the queue
protocol AND the run-event vocabulary over one pg driver (design "Library
Choices"), extract the writable agent-DB data layer into a new @mymemo/agent-db
package: the Drizzle schema, the Database client, the run-store transaction
helpers, the RunEventType vocabulary, the PGlite test harness, and the
migrations. chat-api's db/*, run-store.ts, and run-event-types.ts re-export
from it; run admission (createQueuedRunStartedTx, PostgresRunStore) and the
client-frame projector stay in chat-api. The worker writes
RunEventType.AssistantText, which the projector maps to the text_delta frame --
sharing the vocabulary is what keeps writer and projector from desyncing.

Cover the split-runtime path with a CI-gated integration test rather than the
Docker-compose harness: e2e/integration.test.ts runs chat-api + agent-worker as
real processes against a real Postgres (a service container -- no image build)
and asserts the streamed SSE frames, wired as an `integration` job in ci.yml so
the worker->projector seam is exercised on every PR (the compose-build e2e never
ran in CI and had rotted to the prototype path). compose.yaml keeps the
agent-worker service as a manual local stack and for the M7 decommission of
gateway/sandbox.
