+++ apps/agent-worker/package.json
@@ -10,11 +10,14 @@
 		"check": "biome check --write"
 	},
 	"dependencies": {
+		"@mymemo/agent-db": "workspace:*",
+		"drizzle-orm": "^0.45.2",
 		"pg": "^8.22.0",
 		"pino": "^9.14.0"
 	},
 	"devDependencies": {
 		"@biomejs/biome": "^2.3.11",
+		"@electric-sql/pglite": "^0.5.3",
 		"@types/bun": "^1.3.11",
 		"@types/pg": "^8.20.0",
 		"e2b": "^2.14.0"
