+++ apps/chat-api/Dockerfile
@@ -28,6 +28,7 @@ FROM base AS prerelease
 COPY --from=install /temp/dev/node_modules node_modules
 COPY apps/chat-api/ apps/chat-api/
 COPY apps/sandbox-daemon/ apps/sandbox-daemon/
+COPY packages/agent-db/ packages/agent-db/
 COPY packages/llm-token/ packages/llm-token/
 COPY package.json .
 
@@ -41,11 +42,13 @@ COPY --from=install /temp/prod/node_modules node_modules
 COPY --from=prerelease /usr/src/app/apps/chat-api/src apps/chat-api/src
 COPY --from=prerelease /usr/src/app/apps/chat-api/package.json apps/chat-api/
 COPY --from=prerelease /usr/src/app/apps/chat-api/tsconfig.json apps/chat-api/
-# Drizzle migrations for the writable DB (mymemo_agent). Shipped so the image can
-# run `bun run db:migrate` (the compose `migrate` one-shot, or a prod deploy step)
-# — the runner reads SQL from this folder; drizzle-kit is not needed at runtime.
-COPY --from=prerelease /usr/src/app/apps/chat-api/drizzle apps/chat-api/drizzle
 COPY --from=prerelease /usr/src/app/apps/sandbox-daemon/dist apps/sandbox-daemon/dist
+# The shared writable-DB data layer. Ships its `drizzle/` migrations so the image
+# can run `bun run db:migrate` (the compose `migrate` one-shot, or a prod deploy
+# step) — the runner reads SQL from `packages/agent-db/drizzle`; drizzle-kit is
+# not needed at runtime — and its `src/` so chat-api's `@mymemo/agent-db` imports
+# resolve through the workspace symlink in node_modules.
+COPY --from=prerelease /usr/src/app/packages/agent-db packages/agent-db
 COPY --from=prerelease /usr/src/app/packages/llm-token packages/llm-token
 COPY --from=prerelease /usr/src/app/package.json .
