+++ apps/chat-api/package.json
@@ -7,7 +7,6 @@
 		"format": "biome format --write",
 		"lint": "biome lint --write",
 		"check": "biome check --write",
-		"db:generate": "drizzle-kit generate",
 		"db:migrate": "bun run src/db/migrate.ts",
 		"smoke:gateway": "bun run scripts/smoke-gateway.ts",
 		"e2b:build:dev": "cd sandbox-template && bun run build.dev.ts",
@@ -15,6 +14,7 @@
 	},
 	"dependencies": {
 		"@hono/standard-validator": "^0.1.5",
+		"@mymemo/agent-db": "workspace:*",
 		"@mymemo/llm-token": "workspace:*",
 		"@scalar/hono-api-reference": "^0.9.22",
 		"@statsig/statsig-node-core": "^0.19.8",
@@ -31,7 +31,6 @@
 		"@biomejs/biome": "^2.3.11",
 		"@electric-sql/pglite": "^0.5.3",
 		"@types/bun": "^1.3.11",
-		"@types/pg": "^8.20.0",
-		"drizzle-kit": "^0.31.10"
+		"@types/pg": "^8.20.0"
 	}
 }
