+++ apps/chat-api/src/features/run-events/run-event-types.ts
@@ -1,37 +1,25 @@
 /**
- * The split-runtime run-event vocabulary and the client-facing SSE frame shapes.
+ * The client-facing SSE frame shapes and the projector's terminal-type set.
  *
- * This module is the single authority for what a client can see. Internal run
- * events are recorded in the durable `run_events` table by the worker (and, for
- * `run_started`, by chat-api when it queues the run); the projector derives the
- * client SSE stream from them via {@link projectRunEvent}. The mapping is
- * fail-closed: an internal event type without an explicit frame mapping produces
- * no frame, so a new internal event type can never leak to clients by accident.
+ * The `run_events.type` write-side vocabulary ({@link RunEventType}) lives in the
+ * shared `@mymemo/agent-db` package so every appender — chat-api's run queue and
+ * the agent-worker — writes the same strings the projector reads; it is
+ * re-exported here so chat-api keeps importing it from `@/features/run-events`.
+ *
+ * This module is the single authority for what a client can SEE: the projector
+ * ({@link projectRunEvent}) derives the client SSE stream from recorded events,
+ * fail-closed — an internal event type without an explicit frame mapping
+ * produces no frame, so a new internal event type can never leak to clients by
+ * accident.
  *
  * The prototype-era `sandbox_id` / `agent_session_id` frames are deliberately
  * absent: they are internal runtime identifiers, found in `run_events`, not in
  * the client stream (see the design doc's client contract).
  */
 
-/**
- * The `run_events.type` values the projector understands. Appenders (chat-api's
- * run queue, the worker) must import these constants rather than hard-code the
- * strings, so this module stays the one place the vocabulary is defined.
- */
-export const RunEventType = {
-	/** First event of a run. Payload `{ conversationId, runId }`. */
-	Started: "run_started",
-	/** A chunk of streamed assistant text. Payload `{ text }`. */
-	AssistantText: "assistant_text",
-	/** Terminal: the run finished successfully. */
-	Done: "run_done",
-	/** Terminal: the run failed. Payload `{ message }`. */
-	Error: "run_error",
-	/** Terminal: the run was canceled by the user. */
-	Canceled: "run_canceled",
-} as const;
+import { RunEventType } from "@mymemo/agent-db/run-events";
 
-export type RunEventType = (typeof RunEventType)[keyof typeof RunEventType];
+export { RunEventType };
 
 /**
  * The terminal event types. When the projector reads one of these it emits the
