+++ biome.json
@@ -5,6 +5,9 @@
 		"clientKind": "git",
 		"useIgnoreFile": true
 	},
+	"files": {
+		"includes": ["**", "!**/drizzle/**"]
+	},
 	"formatter": {
 		"enabled": true,
 		"indentStyle": "tab"
