+++ packages/agent-db/package.json
@@ -1,0 +1,34 @@
+{
+	"name": "@mymemo/agent-db",
+	"version": "0.1.0",
+	"type": "module",
+	"exports": {
+		".": "./src/index.ts",
+		"./schema": "./src/schema.ts",
+		"./client": "./src/client.ts",
+		"./run-store": "./src/run-store.ts",
+		"./run-events": "./src/run-events.ts",
+		"./testing": "./src/testing.ts",
+		"./migrations": "./src/migrations.ts"
+	},
+	"scripts": {
+		"test": "bun test",
+		"db:generate": "drizzle-kit generate",
+		"format": "biome format --write",
+		"lint": "biome lint --write",
+		"check": "biome check --write"
+	},
+	"peerDependencies": {
+		"drizzle-orm": "^0.45.2",
+		"pg": "^8.22.0"
+	},
+	"devDependencies": {
+		"@biomejs/biome": "^2.3.11",
+		"@electric-sql/pglite": "^0.5.3",
+		"@types/bun": "^1.3.11",
+		"@types/pg": "^8.20.0",
+		"drizzle-kit": "^0.31.10",
+		"drizzle-orm": "^0.45.2",
+		"pg": "^8.22.0"
+	}
+}
