+++ packages/agent-db/src/index.ts
@@ -1,0 +1,9 @@
+// Barrel for the writable agent-DB data layer. Note `./testing` is deliberately
+// NOT re-exported here: it imports pglite (a devDependency), which is absent
+// from production installs, so pulling it into the root barrel would break
+// production imports. Import the test harness from `@mymemo/agent-db/testing`.
+export * from "./client";
+export * from "./migrations";
+export * from "./run-events";
+export * from "./run-store";
+export * from "./schema";
